post man is add, exception(uses help from net) file is added

+++ mid_term/src/main/java/org/demo/service/CourseServiceImpl.java
@@ -1,6 +1,8 @@
 package org.demo.service;
 
 import org.demo.entity.CourseEntity;
+import org.demo.exception.CourseNotFoundException;
+import org.demo.exception.CourseNotFoundInCartException;
 import org.demo.repo.CourseRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,23 +24,28 @@ public class CourseServiceImpl implements CourseService {
     @Override
     public CourseEntity addCourse(CourseEntity courseEntity) {
         System.err.println("Starting to add course...");
-
+        courseEntity.corseGoingToAdded();
         return courseRepo.save(courseEntity);
     }
 
     @Override
     public CourseEntity getCourseDetails(Long id) {
-        return courseRepo.findById(id).orElse(null);
+        System.err.println("searching for id : " + id);
+        return courseRepo.findById(id)
+                .orElseThrow(() -> new CourseNotFoundException("Course with ID " + id + " not found."));
     }
 
     @Override
     public List<CourseEntity> getAllCourses() {
+        System.err.println("Giving all the courses");
         return courseRepo.findAll();
     }
 
     @Override
     public CourseEntity updateCourse(Long id, CourseEntity updatedCourse) {
-        CourseEntity courseEntity = courseRepo.getById(id);
+        System.err.println("updating course with id : " + id);
+        CourseEntity courseEntity = courseRepo.findById(id)
+                .orElseThrow(() -> new CourseNotFoundException("Cannot update. Course with ID " + id + " not found."));
         courseEntity.setCourseName(updatedCourse.getCourseName());
         courseEntity.setAuthorName(updatedCourse.getAuthorName());
         courseEntity.setDuration(updatedCourse.getDuration());
@@ -48,11 +55,16 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public void deleteCourse(Long id) {
+        System.err.println("deleting course with id : " + id);
+        if (!courseRepo.existsById(id)) {
+            throw new CourseNotFoundException("Cannot delete. Course with ID " + id + " does not exist.");
+        }
         courseRepo.deleteById(id);
     }
 
     @Override
     public void deleteAllCourses() {
+        System.err.println("deleting all courses");
         courseRepo.deleteAll();
     }
 
@@ -68,13 +80,17 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public Map<Long, CourseEntity> addToCart(Long courseId) {
-        CourseEntity courseEntity=courseRepo.findById(courseId).get();
-        CourseEntity.cart.put(courseId,courseEntity);
+        CourseEntity courseEntity = courseRepo.findById(courseId)
+                .orElseThrow(() -> new CourseNotFoundException("Cannot add to cart. Course with ID " + courseId + " not found."));
+        CourseEntity.cart.put(courseId, courseEntity);
         return CourseEntity.cart;
     }
 
     @Override
     public Map<Long, CourseEntity> removeFromCart(Long courseId) {
+        if (!CourseEntity.cart.containsKey(courseId)) {
+            throw new CourseNotFoundInCartException("Cannot remove. Course with ID " + courseId + " is not in the cart.");
+        }
         CourseEntity.cart.remove(courseId);
         return CourseEntity.cart;
     }
